fix(client): resolve circular dependency in shared tabOrder plumbing
The standalone WidgetProvider/factory/tabOrderPropertyConfig.ts imported
constants/PropertyControlConstants and was imported by the WidgetFactory,
closing a new import cycle that the ci-client-cyclic-deps-check (dpdm) gate
flagged as +1 over the base branch.

Move the shared tabOrder property-pane helpers into the existing
WidgetProvider/factory/helpers.ts, which already imports every module they
need (PropertyControlConstants, WidgetValidation, ./types,
widgets/wds/constants) and is already part of these factory cycles. This
adds no new import edge, so the dpdm circular-dependency count returns to
the base 2134 with zero tabOrder-related cycles.

- Delete tabOrderPropertyConfig.ts; functions/constants now live in helpers.ts
- factory/index.tsx imports addTabOrderToPropertyPaneConfig from ./helpers
- Rename the config test to helpers.tabOrder.test.ts, importing from ./helpers

No behavior change.
